pyproject.toml
[project]
name = "message-qa-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.3",
    "langchain-openai>=1.0.2",
    "pymupdf>=1.26.6",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
